Fix bug 1009801 - Properly handle events that fail validation
If you attempt to create an event that has a malformed registerLink,
the server will throw:

Possibly unhandled TypeError: Cannot call method 'setTags' of undefined
    at db.event.create.then.then.then.then.res.json.message
    (webmaker-events-service/routes/dbController.js:234:27)

This is because Sequelize attempts to execute all of the chained
Promises after the initial event save promise is rejected. The proper
way to handle this is to add a .catch() promise to the end of the
chained promise.